switches statsd dependency to equivalent commit referenced by temporal server
Fixes an issue where an ambiguous import results when trying to switch the reference from uber's ringpop to temporal's fork of ringpop. We fix this by making sure that both temporal's ringpop and temporal's server are aligned on the same dependency. This ends up effectively upgrading this repo's statsd reference from 3.1.0 to 3.2.0 (which is what temporal is referencing as well)